etl/requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pymongo==4.7.3

